WIP refactoring, lots of damage
